Rework endpoint cancellation code
Since Python 3.8, CancelledError is no longer a subclass of Exception,
so the separate cancellation handling in the senders isn't strictly
necessary.

If gather() isn't passed return_exceptions=True, then it will raise a
CancelledError (because the things it's gathering were cancelled), thus
ending the endpoint's thread of execution early. But that traceback
never gets printed anywhere, which was confusing; better to explicitly
discard tracebacks in the individual endpoints.

Possibly the final call to gather() in main() should also use
return_exceptions=True. The way it is now, it will return at the first
exception, even if other tasks are still running. However, that is
probably actually the behaviour we want, or close enough, since aiorun
will cancel the other tasks for us.

This is all messy and horrible and I hate it.